requirements.txt
flask==3.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
uuid==1.30
#Python 3.8+
